Move resolveLabwareLocation() to its own file.
